requirements.txt
Flask==2.0.1
requests==2.26.0
beautifulsoup4==4.9.3
Pillow==8.3.1
